On branch default-main 	modified:   .gitignore 	modified:   code/backend/queries/EventOps.cts 	modified:   code/backend/queries/InvitationOps.cts 	modified:   code/backend/queries/MesOps.cts 	modified:   code/backend/queries/NotifOps.cts 	modified:   code/backend/queries/RequestOps.cts 	modified:   code/backend/queries/SessionOps.cts 	modified:   code/backend/queries/UserOps.cts 	modified:   code/backend/queries/logging.cts 	modified:   code/backend/server/ShutDown.cts 	renamed:    code/backend/server/pairs.cts -> code/backend/server/data.cts 	modified:   code/backend/server/db.cts 	modified:   code/backend/server/endpoints.cts 	modified:   code/backend/server/main.cts 	modified:   package.json 	deleted:    tsconfig.json - Removed unecessary depedencies. - Renamed `pairs.cts`. - Removed unecessary commas.
