+++ .gitignore
@@ -2,4 +2,4 @@
 #*#
 node_modules/
 yarn.lock
-prod/
+package-lock.json
