+++ code/backend/queries/InvitationOps.cts
@@ -1,17 +1,17 @@
 const g_coRouter = require("express").Router()
-const g_csRoot = require("../utilities/root.cjs")
-const g_coInvitations = require("../server/main.cjs").get("DB").collection("invitations")
-const g_codes = require("../server/pairs.cjs").get("Status codes")
+const g_csRoot = require("../utilities/root.cts")
+const g_coInvitations = require("../server/main.cts").get("DB").collection("invitations")
+const g_codes = require("../server/data.cts").get("Status codes")
 
 g_coRouter.post(g_csRoot, function(a_oRequest, a_oResponse) {
 	
-}),
+})
 g_coRouter.get(g_csRoot, function(a_oRequest, a_oResponse) {
 	
-}),
+})
 g_coRouter.put(g_csRoot, function(a_oRequest, a_oResponse) {
 	
-}),
+})
 g_coRouter.delete(g_csRoot, function(a_oRequest, a_oResponse) {
 	
 })
