+++ code/backend/server/main.cts
@@ -1,6 +1,6 @@
 process.stdout.write("Hello world!\n")
 require("dotenv").config()
-const g_cToggleProcessing = require("../utilities/processing.cjs")
+const g_cToggleProcessing = require("../utilities/processing.cts")
 
 let g_bShuttingDown = false
 JSON.parse(process.env.SIGNALS).forEach(a_signal => process.on(a_signal, async function() {
@@ -8,7 +8,7 @@ JSON.parse(process.env.SIGNALS).forEach(a_signal => process.on(a_signal, async f
 	if (!g_bShuttingDown) {
 		g_bShuttingDown = true
 		g_cToggleProcessing()
-		for (const l_cFunction of require("./ShutDown.cjs")) {
+		for (const l_cFunction of require("./ShutDown.cts")) {
 			await l_cFunction()
 		}
 		process.exit(Number(process.stdout.write("Absolute cinema\n")))
@@ -24,14 +24,14 @@ module.exports = g_coEssentials;
 
 (async function() {
 	try {
-		g_coEssentials.set("DB", await require("./db.cjs")())
+		g_coEssentials.set("DB", await require("./db.cts")())
 	} catch (a_oError) {
 		g_cToggleProcessing()
 		console.error("Unable to connect to database due to: ", a_oError)
 		return process.kill(process.pid)
 	}
 	g_cToggleProcessing("Attempting server start.")
-	g_coEssentials.set("Server", require("./endpoints.cjs")
+	g_coEssentials.set("Server", require("./endpoints.cts")
 		.listen(process.env.PORT, function() {
 			g_cToggleProcessing()
 			console.log("Server started.")
