+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "GroupProject",
   "version": "1.0.0",
-  "main": "prod/backend/server/main.cjs",
+  "main": "code/backend/server/main.cts",
   "repository": "https://github.com/Group-tam/full-stack-development",
   "author": "Gr 8",
   "license": "MIT",
@@ -11,9 +11,5 @@
     "express": "latest",
     "express-session": "latest",
     "mongodb": "latest"
-  },
-  "devDependencies": {
-    "@types/node": "latest",
-    "typescript": "latest"
   }
 }
