--- tsconfig.json
@@ -1,10 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ESNext",
-        "module": "NodeNext",
-        "moduleResolution": "NodeNext",
-        "outDir": "prod/backend",
-        "esModuleInterop": false
-    },
-    "include": ["code/backend/**/*.cts"]
-}
